feature(application): Introduces \Elgg\Application object and moved most of the booting process into it.
Deprecated inclusion of start.php for booting the core.

Refs #7721
