+++ engine/tests/phpunit/bootstrap.php
@@ -15,7 +15,16 @@ $CONFIG = (object) array(
 	'dbprefix' => 'elgg_',
 	'boot_complete' => false,
 	'wwwroot' => 'http://localhost/',
+	'dataroot' => __DIR__ . '/test_files/dataroot/',
 	'site_guid' => 1,
 );
 
-require_once "$engine/load.php";
+$autoload_path = dirname(dirname(dirname(dirname(__FILE__)))) . '/vendor/autoload.php';
+$autoload_available = include_once($autoload_path);
+if (!$autoload_available) {
+	die("Couldn't include '$autoload_path'. Did you run `composer install`?");
+}
+
+$app = new \Elgg\Application();
+
+$app->loadCore();
